Updated DexDividend.
